+++ pkg/_macros/lib/src/executor/augmentation_library.dart
@@ -145,6 +145,7 @@ class _Builder {
       {List<Span>? spans}) {
     Map<Identifier, List<(Key, DeclarationCode)>> mergedTypeResults = {};
     Map<Identifier, List<(Key, DeclarationCode)>> mergedEntryResults = {};
+    Map<Identifier, (Key, NamedTypeAnnotationCode)> mergedExtendsResults = {};
     Map<Identifier, List<(Key, TypeAnnotationCode)>> mergedInterfaceResults =
         {};
     Map<Identifier, List<(Key, TypeAnnotationCode)>> mergedMixinResults = {};
@@ -164,6 +165,15 @@ class _Builder {
         mergedEntryResults.update(
             identifier, (enumValues) => enumValues..addAll(values),
             ifAbsent: () => values.toList());
+      });
+      result.extendsTypeAugmentations.forEach((identifier, value) {
+        mergedExtendsResults.update(
+            identifier,
+            (existing) => throw StateError(
+                'A class cannot extend multiple classes, ${identifier.name} '
+                'tried to extend both ${existing.$2.name.name} and '
+                '${value.name.name}.'),
+            ifAbsent: () => (IdentifierKey.superclass(key, identifier), value));
       });
       result.interfaceAugmentations.forEach((identifier, value) {
         int index = 0;
@@ -224,6 +234,14 @@ class _Builder {
       if (keywords.isNotEmpty) keywords.add('');
       _writeDirectiveStringPart(TypeDeclarationContentKey.declaration(key),
           'augment ${keywords.join(' ')}$declarationKind ${type.name} ');
+
+      if (mergedExtendsResults[type] case (var superclassKey, var superclass)) {
+        Key fixedKey = TypeDeclarationContentKey.superclass(key);
+        int index = 0;
+        _buildString(fixedKey, index++, 'extends ');
+        _buildCode(superclassKey, superclass);
+        _buildString(fixedKey, index++, ' ');
+      }
 
       if (mergedMixinResults[type] case var mixins? when mixins.isNotEmpty) {
         Key mixinsKey = TypeDeclarationContentKey.mixins(key);