+++ pkg/_macros/test/executor/simple_macro.dart
@@ -6,6 +6,8 @@ import 'dart:async';
 
 import 'package:_macros/src/api.dart';
 
+import '../util.dart';
+
 /// A macro for testing diagnostics reporting, including error handling.
 class DiagnosticMacro implements ClassTypesMacro {
   @override
@@ -467,7 +469,7 @@ class SimpleMacro
 
   @override
   FutureOr<void> buildTypesForClass(
-      ClassDeclaration clazz, ClassTypeBuilder builder) {
+      ClassDeclaration clazz, ClassTypeBuilder builder) async {
     List<Object> buildTypeParam(
         TypeParameterDeclaration typeParam, bool isFirst) {
       return [
@@ -516,6 +518,10 @@ mixin $mixinName implements $interfaceName {
   int get x => 1;
 }'''));
 
+    // ignore: deprecated_member_use_from_same_package
+    final mySuperClass = await builder.resolveIdentifier(
+        Fixtures.library.uri, Fixtures.mySuperclass.identifier.name);
+    builder.extendsType(NamedTypeAnnotationCode(name: mySuperClass));
     builder.appendInterfaces([RawTypeAnnotationCode.fromString(interfaceName)]);
     builder.appendMixins([RawTypeAnnotationCode.fromString(mixinName)]);
   }