+++ pkg/macros/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0-main.6
+
+- Add extendsType API for adding an extends clause.
+- Refactor builder implementations, fixes some bugs around nested builders.
+
 ## 0.1.0-main.5
 
 - Handle ParallelWaitError with DiagnosticException errors nicely.